updated simulator
